+++ packages/grafana-ui/src/components/DataLinks/DataLinkInput.tsx
@@ -1,9 +1,8 @@
 import { lazy, memo, Suspense } from 'react';
 
 import { type VariableSuggestion } from '@grafana/data';
-import { t } from '@grafana/i18n';
 
-import { LoadingPlaceholder } from '../LoadingPlaceholder/LoadingPlaceholder';
+import { Input } from '../Input/Input';
 
 export type DataLinkInterpolationMode = 'url' | 'text';
 
@@ -35,8 +34,27 @@ const DataLinkInputImplementation = lazy(() =>
 
 export const DataLinkInput = memo(function DataLinkInput(props: DataLinkInputProps) {
   return (
-    <Suspense fallback={<LoadingPlaceholder text={t('grafana-ui.data-link-input.loading', 'Loading editor')} />}>
+    <Suspense fallback={<DataLinkInputLoadingFallback {...props} />}>
       <DataLinkInputImplementation {...props} />
     </Suspense>
   );
 });
+
+function DataLinkInputLoadingFallback({
+  value,
+  placeholder = 'http://your-grafana.com/d/000000010/annotations',
+  id = 'data-link-input',
+}: DataLinkInputProps) {
+  return (
+    <Input
+      id={`${id}-loading`}
+      value={value}
+      placeholder={placeholder}
+      readOnly
+      tabIndex={-1}
+      aria-label=""
+      aria-labelledby=""
+      aria-hidden
+    />
+  );
+}
