--- public/app/features/notebook/scene/layout-notebook/cells/codeCellFocusExtension.ts
@@ -1,16 +1,0 @@
-import { ViewPlugin } from '@codemirror/view';
-
-import { type CodeMirrorExtension } from '@grafana/ui/unstable';
-
-/**
- * Builds a fresh extension that moves focus into a CodeMirror editor after its
- * view has been attached to the DOM.
- */
-export function buildCodeCellFocusExtension(): CodeMirrorExtension[] {
-  return [
-    ViewPlugin.define((view) => {
-      requestAnimationFrame(() => view.focus());
-      return {};
-    }),
-  ];
-}
